backend/requirements/base.txt
Django>=4.2,<5.0
djangorestframework>=3.14
djangorestframework-simplejwt>=5.3
django-cors-headers>=4.3
django-filter>=23.5
django-environ>=0.11
dj-database-url>=2.1
psycopg2-binary>=2.9
cloudinary>=1.36
django-cloudinary-storage>=0.3.0
Pillow>=10.0
whitenoise>=6.6
gunicorn>=21.2
